Rewrite in Vue
This commits rewrites the whole frontend in Vue to be able to better
support new features and a better structure for the future.

Features should be mostly identical, but the overall look should be a
liitle bit more consistent and a little bit nicer. Also the reload
should work a little bit better.

Closes #20