Replay missed Database Changes back to the client
With this commit an `id` is introduced in each SSE, which is just the
Unix Timestamp in Milliseonds from the created, updated or deleted
object from the database. On reconnection the client sends the
timestamps as a header `Last-Event-ID` and the server can then replay
all changes after this timestamp again to the client. This way the
client should also be up to date after a dis- and re-connect happend.

To make the time even shorter and to circumvent the SSE Connection to
close, this commit introduces a ping SSE, which is sent on a regular
base.