Typifier Service
With this commit the listinator becomes a multi service project. It
introduces another service written in python which can predict the type
on a product you want to buy. It uses simple ML tools and currently a
very small trainings set. There is a lot room for improvement, but it
already can predict the types quite good.

Since this commit splits the repository in several services, the
struckture is changed and there are now Makefiles for each part.